+++ src/register.js
@@ -1,3 +1,0 @@
-'use strict';
-
-require('./dist/index.js').register(window.jQuery || window.$);